Move Authenticator to separate file for easier merging with closed-source
